refactor #step1 Parameters만 일급컬렉으로 구현. HttpURL 삭제 처리
