+++ src/main/java/http/RequestLine.java
@@ -6,21 +6,35 @@ public class RequestLine {
   private static final int METHOD_INDEX = 0;
   private static final int URL_INDEX = 1;
   private static final int VERSION_INDEX = 2;
+  private static final String QUESTION_DELIMITER = "\\?";
+  private static final int PATH_INDEX = 0;
+  private static final int QUERYSTRING_INDEX = 1;
+  private static final int HAS_QUERYSTRING_CONDITION_LENGTH = 1;
 
   private String method;
-  private HttpURL url;
+  private String path;
   private String version;
+  private Parameters parameters;
 
-  private RequestLine(String method, String url, String version) {
+
+  private RequestLine(String method, String path, String version, Parameters parameters) {
     this.method = method;
-    this.url = HttpURL.parse(url);
+    this.path = path;
     this.version = version;
+    this.parameters = parameters;
   }
 
   public static RequestLine parse(String requestLine) {
     String[] requestLineToken = requestLine.split(SPACE_DELIMITER);
-    return new RequestLine(requestLineToken[METHOD_INDEX], requestLineToken[URL_INDEX],
-        requestLineToken[VERSION_INDEX]);
+    String[] urlToken = requestLineToken[URL_INDEX].split(QUESTION_DELIMITER);
+
+    return new RequestLine(requestLineToken[METHOD_INDEX], urlToken[PATH_INDEX],
+        requestLineToken[VERSION_INDEX], makeParameters(urlToken));
+  }
+
+  private static Parameters makeParameters(String[] urlToken) {
+    return urlToken.length > HAS_QUERYSTRING_CONDITION_LENGTH ? Parameters
+        .parse(urlToken[QUERYSTRING_INDEX]) : Parameters.EMPTY;
   }
 
   public String getMethod() {
@@ -32,10 +46,10 @@ public class RequestLine {
   }
 
   public String getPath() {
-    return url.getPath();
+    return path;
   }
 
-  public String getParameter(String key) {
-    return url.getParameter(key);
+  public Parameters getParameters() {
+    return parameters;
   }
 }
