+++ src/test/java/http/RequestLineTest.java
@@ -13,9 +13,24 @@ public class RequestLineTest {
     RequestLine requestLine = RequestLine
         .parse("GET /users?userId=javajigi&password=password&name=JaeSung&noValue= HTTP/1.1");
 
+    Parameters parameters = Parameters.parse("userId=javajigi&password=password&name=JaeSung&noValue=");
+
     assertThat(requestLine.getMethod()).isEqualTo("GET");
     assertThat(requestLine.getPath()).isEqualTo("/users");
     assertThat(requestLine.getVersion()).isEqualTo("HTTP/1.1");
+    assertThat(requestLine.getParameters()).isEqualTo(parameters);
+  }
+
+  @Test
+  @DisplayName("queryString이 없을때 Parameters는 빈값이다")
+  void emptyQueryString() {
+    RequestLine requestLine = RequestLine
+        .parse("GET /users HTTP/1.1");
+
+    assertThat(requestLine.getMethod()).isEqualTo("GET");
+    assertThat(requestLine.getPath()).isEqualTo("/users");
+    assertThat(requestLine.getVersion()).isEqualTo("HTTP/1.1");
+    assertThat(requestLine.getParameters()).isEqualTo(Parameters.EMPTY);
   }
 
 }
